+++ SciSDK_VC++/SciSDK_VC++/SciSDK_VC++.cpp
@@ -265,7 +265,7 @@ void DemoV2495_V4718_ETH() {
 	uint32_t v;
 
 	//TEST SCOPE
-	res = sdk.AddNewDevice("V4718USB:59733:0x32100000", "V2495", "RegisterFileV2495.json", "board0");
+	res = sdk.AddNewDevice("V4718ETH:V4718-USB-59733.local:0x32100000", "V2495", "RegisterFileV2495.json", "board0");
 	sdk.SetParameter("board0:/boardapi/gd[0].delay", 100);
 	sdk.p_error(sdk.SetRegister("board0:/Registers/bb", 1));
 	sdk.p_error(sdk.GetRegister("board0:/Registers/aa", &v));
@@ -418,13 +418,13 @@ int main()
 
 	//Just run one demo at time!
 
-	//OpenDppDemo();
+	OpenDppDemo();
 	//OpenScopeDemo();
 	//DemoSciCompilerOscilloscope();
 	//DemoV2495_USB_LogicAnalyser();
 	//DemoV2495_USB_RegisterPerformance();
 	//DemoSciCompilerList2();
-	DemoV2495_V4718_ETH();
+	//DemoV2495_V4718_ETH
 	return 0;
 
 }